refactor(runner-core): split redaction tests
